+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App'
+import App from './Components/App/App'
 import './index.css';
-import './App.css'
+import './Components/App/App.css'
 
 
 const root = document.getElementById('root');
